requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.9.2
azure-ai-agents==1.1.0
azure-ai-projects==1.0.0
azure-identity==1.25.2
azure-core==1.38.2
openai==1.68.2
pyodbc==5.0.1
requests==2.32.3
python-dotenv==1.0.0
typing-extensions==4.12.2
azure-keyvault-secrets
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.15
azure-ai-projects
opentelemetry-sdk
azure-monitor-opentelemetry
azure-ai-contentsafety
